pyproject.toml
[project]
name = "nogfi-hackathon"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "t1rasane", email = "timo.rasanen@nokia.com" }
]
dependencies = [
    "napalm~=5.0.0",
    "nornir~=3.4.1",
    "pynetbox~=7.4.0",
    "fakenos~=1.0.1",
    "pydantic~=2.8.2",
    "pydantic-settings~=2.4.0",
    "scrapli[asyncssh]~=2024.7.30",
    "ansible~=10.3.0",
    "genie~=24.7",
    "pyats~=24.7",
    "napalm-h3c-comware~=0.1.5",
    "napalm-ftos @ git+https://github.com/napalm-automation-community/napalm-ftos",
    "napalm-ros @ git+https://github.com/napalm-automation-community/napalm-ros",
    "napalm-arubaos-switch~=0.2.0",
    "napalm-huawei-vrp~=1.2.0",
    "scrapli-netconf~=2024.7.30",
    "nornir-napalm~=0.5.0",
    "nornir-netmiko~=1.0.1",
    "httpx~=0.27.2",
    "loguru~=0.7.2",
    "click~=8.1.7",
    "english-words~=2.0.1",
    "pygnmi~=0.8.14",
    "nornir-pygnmi~=0.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"nogfi-hackathon" = "nogfi_hackathon:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.rye]
managed = true
dev-dependencies = []
